lianxiang-0727

+++ lianxiang_0727/src/main/java/huawei_t/t6.java
@@ -8,7 +8,7 @@ import java.util.Scanner;
 public class t6 {
 
 
-    static int[][][] find ;
+    static long[][][] find ;
     static boolean[][][] visited ;
 //    static int T, n, m, k;
 
@@ -18,11 +18,11 @@ public class t6 {
         int n = sc.nextInt();
         int m = sc.nextInt();
         int k = sc.nextInt();
-        find = new int[n+1][m+1][k+1];
+        find = new long[n+1][m+1][k+1];
         for (int i = 0; i < find.length; i++) {
             for (int j = 0; j < find[0].length; j++) {
                 for (int l = 0; l < find[0][0].length; l++) {
-                    find[i][j][k] = Integer.MAX_VALUE;
+                    find[i][j][k] = Long.MAX_VALUE;
                 }
             }
         }
@@ -30,17 +30,16 @@ public class t6 {
         System.out.println(dfs(n, m, k));
     }
 
-    public static int dfs(int n, int m, int k) {
+    public static long dfs(long n, long m, long k) {
         if (n > m) return dfs(m, n, k);
         if (visited[n][m][k]) return find[n][m][k];     // n*m 划分 k 块需要的
         visited[n][m][k] = true;
-//        if (k == 0 || n * m == k) return find[n][m][k] = 0;
         if (k==1) {
             find[n][m][k] = 0;
             return 0;
         }
         if(k==2){
-            int min = Math.min(m,n);
+            long min = Math.min(m,n);
             find[n][m][k] = min*min;
             return min*min;
         }
